API

Gutzwiller.AbstractAnsatz

Gutzwiller.AnsatzSampling

Gutzwiller.CombinationAnsatz

Gutzwiller.CombinedBlockingResult

Gutzwiller.DensityProfileAnsatz

Gutzwiller.ExtendedGutzwillerAnsatz

Gutzwiller.GradientDescentResult

Gutzwiller.GrossPitaevskiiAnsatz

Gutzwiller.GutzwillerAnsatz

Gutzwiller.JastrowAnsatz

Gutzwiller.KineticVQMC

Gutzwiller.KineticVQMCResult

Gutzwiller.KineticVQMCResultRows

Gutzwiller.KineticVQMCWalkerState

Gutzwiller.LocalEnergyEvaluator

Gutzwiller.MultinomialAnsatz

Gutzwiller.RelativeJastrowAnsatz

Gutzwiller.VectorAnsatz

Gutzwiller.adaptive_gradient_descent

Gutzwiller.amsgrad

Gutzwiller.gradient_descent

Gutzwiller.kinetic_sample!

Gutzwiller.kinetic_vqmc

Gutzwiller.kinetic_vqmc!

Gutzwiller.local_energy_estimator

Gutzwiller.pick_random_from_cumsum

Gutzwiller.resample

Gutzwiller.resample!

Gutzwiller.val_and_grad

Gutzwiller.val_err_and_grad

Gutzwiller.AbstractAnsatz — Type

abstract type AbstractAnsatz{K,V,N}

An ansatz behaves similar to an AbstractDVec with keytype K and valtype V with N parameters.

It must provide the following:

ansatz(key::K, params)::V: Get the value of the ansatz for a given address key with specified parameters.

val_and_grad(ansatz, key, params): Get the value and gradient (w.r.t. the parameters) of the ansatz.

build_basis: for collecting the vector to DVec/PDVec (optional).

Gutzwiller.AnsatzSampling — Type

struct AnsatzSampling(hamiltonian, ansatz, params)

Importance-sampled Hamiltonian, where the importance sampling is controlled by ansatz and params.

julia> H = HubbardRealSpace(BoseFS(1,1), u=6);

julia> ansatz = GutzwillerAnsatz(H);

julia> G = AnsatzSampling(H, ansatz, 0.6);

julia> Matrix(H)
3×3 Matrix{Float64}:
  0.0      -2.82843  -2.82843
 -2.82843   6.0       0.0
 -2.82843   0.0       6.0

julia> Matrix(G)
3×3 Matrix{Float64}:
  0.0        -103.515  -103.515
 -0.0772832     6.0       0.0
 -0.0772832     0.0       6.0

Gutzwiller.CombinationAnsatz — Type

CombinationAnsatz

Superposition of ansatz1 and ansatz2. Has one more parameter than each ansatz controling the mixing between them.

Gutzwiller.CombinedBlockingResult — Type

CombinedBlockingResult{T}

Combination of several independent blocking results.

Gutzwiller.DensityProfileAnsatz — Type

DensityProfileAnsatz(hamiltonian) <: AbstractAnsatz

\[D(|f⟩; p) = exp(-∑_{i=1}^M p_i ⟨f|n_i|f⟩)\]

Gutzwiller.ExtendedGutzwillerAnsatz — Type

ExtendedGutzwillerAnsatz(hamiltonian) <: AbstractAnsatz

The Extended Gutzwiller ansatz:

\[G(|f⟩; 𝐠) = exp(-g_1 ⟨f|H|f⟩ - g_2 ⟨f| ∑_{<i,j>} n_i n_j |f⟩),\]

where $H$ is an ExtendedHubbardReal1D Hamiltonian. The additional term accounts for the strength of nearest-neighbour interactions.

It takes two parameters, g_1 and g_2.

Gutzwiller.GradientDescentResult — Type

GradientDescentResult

Return type of adaptive_gradient_descent variants such as gradient_descent and amsgrad.

Can be used as a Table from Tables.jl.

Gutzwiller.GrossPitaevskiiAnsatz — Type

GrossPitaevskiiAnsatz(h::AbstractHamiltonian)
GrossPitaevskiiAnsatz(address::BoseFS; valtype::Type = Float64) <: AbstractAnsatz

Ansatz representing a Gross-Pitaevskii product state of $N$ bosons in $M$ modes, with single-particle orbital parameter vector $𝐜 = (c_1, …, c_M)$:

\[|Ψ_{\mathrm{GP}}⟩ = \frac{1}{\sqrt{N!}} \left( ∑_{m=1}^M cₘ aₘ^† \right)^N |0⟩\]

Projected onto a Fock basis state $|n_1, …, n_M⟩$, the amplitude is:

\[⟨n₁, …, n_M | Ψ_{\mathrm{GP}}⟩ = \sqrt{\frac{N!}{\prod_{m=1}^M nₘ!}} \prod_{m=1}^M cₘ^{nₘ}\]

Example

julia> gpa = GrossPitaevskiiAnsatz(BoseFS(2, 0))
GrossPitaevskiiAnsatz(BoseFS(2, 0); valtype=Float64)

julia> amplitude = gpa(BoseFS(1,1), [1.0, 2.0])
2.8284271247461903

Gutzwiller.GutzwillerAnsatz — Type

GutzwillerAnsatz(hamiltonian) <: AbstractAnsatz

The Gutzwiller ansatz:

\[G(|f⟩; g) = exp(-g ⟨f|H|f⟩),\]

where $H$ is the hamiltonian passed to the struct.

It takes a single parameter, g.

Gutzwiller.JastrowAnsatz — Type

JastrowAnsatz(hamiltonian) <: AbstractAnsatz

\[J(|f⟩; 𝐩) = exp(-∑_{k=1}^M ∑_{l=k}^M p_{k,l} ⟨f| n_k n_l |f⟩)\]

With translationally invariant Hamiltonians, use RelativeJastrowAnsatz instead.

Gutzwiller.KineticVQMC — Type

KineticVQMC(
    hamiltonian, ansatz;
    samples=1e6,
    walkers=Threads.nthreads() > 1 ? Threads.nthreads() * 2 : 1,
    steps=round(Int, samples / walkers),
)

Usage

julia> ham = HubbardReal1D(BoseFS((1,1,1,1)));

julia> qmc = KineticVQMC(ham, GutzwillerAnsatz(ham); samples=1e5);

julia> qmc(0.5)
-6.650547790454165

julia> val_and_grad(qmc, 0.5)
(-6.6525129909164304, [-0.04590881240854946])

The result of a run is stored in the struct and can be accessed and manipulated after a computation. If qmc or val_and_grad is called again, the result is overwritten.

julia> qmc.result
KineticVQMCResult
  walkers:      24
  samples:      100008
  local energy: -6.6525 ± 0.001509

julia> local_energy_estimator(qmc.result)
CombinedBlockingResult{Float64}
  mean = -6.6525 ± 0.0015
  with uncertainty of ± 4.082709649722105e-5
  Combined from 24 blocking results. (k ∈ 2 … 6)

julia> val_and_grad(qmc.result)
(-6.6525129909164304, [-0.04590881240854946])

Gutzwiller.KineticVQMCResult — Type

KineticVQMCResult

The result of kinetic_vqmc. Holds the state of each walker. Use local_energy_estimator(result) to get an estimate of the local energy, or PDVec(result) to materialize the sampled vector.

Supports the Tables.jl interface and a such can be converted to a DataFrame or saved to file with Arrow.jl or CSV.jl.

Gutzwiller.KineticVQMCResultRows — Type

KineticVQMCResultRows

Tables.jl-compatible row iterator for KineticVQMCResult.

Gutzwiller.KineticVQMCWalkerState — Type

mutable struct KineticVQMCWalkerState(hamiltonian, ansatz)

Contains all information needed to perform continuous time variational quantum Monte Carlo.

See also: kinetic_vqmc!, kinetic_vqmc.

Gutzwiller.LocalEnergyEvaluator — Type

LocalEnergyEvaluator(hamiltonian, ansatz; basis=build_basis(hamiltonian))

Wrapper over hamiltonian and ansatz that can be used to efficiently compute local energies. Stores the basis, which can be passed throught the basis keyword argument.

Instances of LocalEnergyEvaluator can be called with parameters to return a local energy. Alternatively, val_and_grad can be used to compute the value and gradient (with respect to the parameters).

julia> H = HubbardReal1D(BoseFS((1,1,1,1,1)));

julia> le = LocalEnergyEvaluator(H, GutzwillerAnsatz(H));

julia> le(0.5)
-8.22150794790878

julia> val_and_grad(le, 0.5)
(-8.22150794790878, [0.2718589892870141])

Extended help

Using with Optim.jl

You can use the gradient provided by this struct by wrapping it in Optim.only_fg!.

julia> H = HubbardReal1D(BoseFS((1,1,1,1,1)));

julia> le = LocalEnergyEvaluator(H, GutzwillerAnsatz(H));

julia> optimize(le, [0.5])
 * Status: success

 * Candidate solution
    Final objective value:     -8.225604e+00

 * Found with
    Algorithm:     Nelder-Mead

 * Convergence measures
    √(Σ(yᵢ-ȳ)²)/n ≤ 1.0e-08

 * Work counters
    Seconds run:   0  (vs limit Inf)
    Iterations:    9
    f(x) calls:    21

julia> optimize(NLSolversBase.only_fg!(le), [0.5])
 * Status: success

 * Candidate solution
    Final objective value:     -8.225604e+00

 * Found with
    Algorithm:     L-BFGS

 * Convergence measures
    |x - x'|               = 1.19e-08 ≰ 0.0e+00
    |x - x'|/|x'|          = 2.53e-08 ≰ 0.0e+00
    |f(x) - f(x')|         = 0.00e+00 ≤ 0.0e+00
    |f(x) - f(x')|/|f(x')| = 0.00e+00 ≤ 0.0e+00
    |g(x)|                 = 2.30e-12 ≤ 1.0e-08

 * Work counters
    Seconds run:   0  (vs limit Inf)
    Iterations:    5
    f(x) calls:    7
    ∇f(x) calls:   7
    ∇f(x)ᵀv calls: 0

Gutzwiller.MultinomialAnsatz — Type

MultinomialAnsatz(H::AbstractHamiltonian; normalize=false)

The multinomial ansatz

\[F(|n_1, n_2, n_3, …⟩; p) = K ( ∏_{i=1}^M 1/n_i )^p.\]

If $p = 1/2$, this is the exact solution to a non-interacting Bose-Hubbard model. If normalize is set to true, a normalization factor of

\[K = ( rac{N}{M^N})^p\]

is applied.

Gutzwiller.RelativeJastrowAnsatz — Type

RelativeJastrowAnsatz(hamiltonian) <: AbstractAnsatz

For a translationally invariant Hamiltonian, this is equivalent to JastrowAnsatz, but has fewer parameters.

\[R(|f⟩; 𝐩) = exp(-∑_{d=0}^{M/2} p_d ∑_{k=1}^M ⟨f| n_k n_{k + d} |f⟩)\]

Gutzwiller.VectorAnsatz — Type

VectorAnsatz(vector::AbstractDVec) <: AbstractAnsatz

Use vector as 0-parameter ansatz.

Gutzwiller.adaptive_gradient_descent — Method

adaptive_gradient_descent(
    φ, ψ, f, p_init;
    maxiter=100, verbose=true, α=0.01,
    early_stop=true,
    grad_tol=√eps(T), param_tol=√eps(T), val_tol=√(eps(T)),
    first_moment_init, second_moment_init, fix_params,
    kwargs...
)

For immediately useful versions of this function, see [gradient_descent] and [amsgrad]. Returns GradientDescentResult.

Arguments

φ, ψ: functions for updating the first and second moment. These are set by the caller ([gradient_descent] and [amsgrad]).

f: the function to be optimized. Must be callable with parameters of the same shape as p_init.

p_init: initial parameter estimate. Alternatively, a GradientDescentResult may be passed which will continue the optimization where the previous computation stopped.

Keyword arguments

maxiter: maximum iterations performed before stopping.

verbose: set to false to disable printing.

α: the learning rate. Larger values converge faster, but are less stable.

grad_tol, param_tol, val_tol: stop if gradient, parameters, or objective function value change less than this amount in a step.

first_moment_init: initial value for the first moment. Defaults to the gradient of f at p_init.

second_moment_init: initial value for the second moment. Defaults to a vector of zeros.

fix_params: a vector or tuple of booleans. If fix_params[i] is true, the i-th parameter will be kept at its initial value.

Gutzwiller.amsgrad — Method

amsgrad(
    f, p0;
    maxiter=100, verbose=true, α=0.01, β1=0.1, β2=0.01,
    grad_tol=√eps(T), param_tol=√eps(T), val_tol=√(eps(T)),
)

AMSGrad on function f with initial parameters p0. α, β1, and β2 are the parameters controlling AMSGrad.

See adaptive_gradient_descent for descriptions of other keyword arguments. Returns GradientDescentResult.

Gutzwiller.gradient_descent — Method

gradient_descent(
    f, p0;
    maxiter=100, verbose=true, α=0.01,
    grad_tol=√eps(T), param_tol=√eps(T), val_tol=√(eps(T)),
)

Vanilla gradient descent on function f with initial parameters p0.

See adaptive_gradient_descent for descriptions of keyword arguments. Returns GradientDescentResult.

Gutzwiller.kinetic_sample! — Method

kinetic_sample!(offdiag_buffer, prob_buffer, hamiltonian, ansatz, curr_addr)

Use the hamiltonian and curr_addr to sample a new address from ansatz, while also computing the local energy contribution of curr_addr. Return a tuple of the new address, the residence time of the current address, and the local energy.

prob_buffer is a Vector{Float64} used as temporary storage.

The algorithm for sampling is taken from [1].

[1]: https://pubs.acs.org/doi/epdf/10.1021/acs.jctc.8b00780

Gutzwiller.kinetic_vqmc! — Method

kinetic_vqmc!(st::KineticVQMCWalkerState; steps)
kinetic_vqmc!(sts::Vector{KineticVQMCWalkerState}; steps)
kinetic_vqmc!(res::KineticVQMCResult; steps)

Continue a kinetic_vqmc computation, peforming steps additional steps (default is the current number of samples in the input).

Returns its input.

Gutzwiller.kinetic_vqmc — Method

kinetic_vqmc(hamiltonian, ansatz, params; steps=1e6, walkers=Threads.nthreads() * 2)

Use continuous time variational quantum Monte Carlo to sample a addresses and local energies from hamiltonian and ansatz. The ansatz must be indexable by addresses accepted by hamiltonian and must support Base.keytype and Base.valtype.

The sampling samples steps steps per walker. The walkers are independent runs that run on separate threads.

The function returns a KineticVQMCResults, which contains the states of all walkers. To continue a run, use continous_time_vqmc!.

See I. Sabzevari and S. Sharma for a detailed description of the algorithm.

Gutzwiller.local_energy_estimator — Method

local_energy_estimator(res::KineticVQMCResultRows; kwargs...)

Return the local energy estimator from the result of a kinetic_vqmc run. Keyword arguments are passed to blocking_analysis. Returns a CombinedBlockingResult.

Gutzwiller.pick_random_from_cumsum — Method

pick_random_from_cumsum(cumsum)

Pick a random index from cumsum, which should contain a cumulative sum (see Base.cumsum) of values proportional to the probabilities with which the index should be picked.

Gutzwiller.resample! — Method

resample!(result, residence_times, values)

In-place version of resample. The length of the output is equal to length(result).

Gutzwiller.resample — Method

resample(residence_times, values; len=length(values))

Resample unevenly spaced timeseries data into a vector of len evenly-spaced values. Here, the time values[i] occupies is equal to residence_times[i]. The resampling is done in such a way that

    mean(resample(xs, ys; len)) == mean(xs, FrequencyWeights(ys))

independent of len.

Using this function allows us to use blocking_analysis on the resampled data.

See also: resample!.

Gutzwiller.val_and_grad — Function

val_and_grad(::AbstractAnsatz, addr, params)

Return ansatz value at addr and its gradient w.r.t. params.

Gutzwiller.val_err_and_grad — Method

val_err_and_grad(args...)

Return the value, its error and gradient. See val_and_grad.